Packet writer improvement parameter can now have different type, only allow integer and enums, same sign added runtimecheck for max value

+++ pol-core/pol/miscmsg.cpp
@@ -402,7 +402,7 @@ namespace Pol {
 		}
 
 		PktHelper::PacketOut<PktOut_98> msgOut;
-		msgOut->WriteFlipped<u16>( static_cast<u16>( 37 ) ); // static length
+		msgOut->WriteFlipped<u16>( 37u ); // static length
 		msgOut->Write<u32>( the_mob->serial_ext );
 		msgOut->Write( the_mob->name().c_str(), 30, false );
 		msgOut.Send( client );
